Remapped inventory screen to manageinventoryscreen and adjusted api method to bring in all data for the time being.
