+++ BeyondSW5E/src/screens/InventoryScreen.js
@@ -1,49 +1,14 @@
-import React, { useState, useEffect } from 'react'
-import { Text, View, StyleSheet, Button } from 'react-native'
-import swapi from '../api/swapi'
-import ItemCardList from '../components/ItemCardList'
+import React from 'react'
+import { Text, View, StyleSheet } from 'react-native'
 
-const InventoryScreen = () => {
-    const [equipment, setEquipment] = useState([])
-
-    const searchApi = async ( limit ) => {
-        const response = await swapi.get('/equipment');
-        const limitArray = []
-        for (let i = 0; i < limit; i++) {
-            limitArray[i] = response.data[i]
-        }
-        setEquipment(limitArray)
-    }
-
-    useEffect(() => {
-        searchApi(20)},
-        [])
-
+const ManageInventoryScreen = () => {
     return (
-        <View style = { styles.screenContainer }>
-            <Text style = { styles.headerStyle }>Inventory Screen</Text>
-            <Text>Displaying {equipment.length} items.</Text>
-            {
-                equipment.length === 0
-                ? <Text>Equipment is Empty</Text>
-                : <ItemCardList
-                      equipment = { equipment }
-                  />
-            }
+        <View>
+            <Text>Manage Inventory Screen</Text>
         </View>
     )
 }
 
-const styles = StyleSheet.create({
-    screenContainer: {
-        alignItems: 'center',
-        flex: 1
-    },
-    headerStyle: {
-        fontSize: 24,
-        fontWeight: 'bold',
-        alignSelf: 'center'
-    }
-})
+const styles = StyleSheet.create({})
 
-export default InventoryScreen
+export default ManageInventoryScreen
