+++ BeyondSW5E/src/screens/ManageInventoryScreen.js
@@ -1,14 +1,51 @@
-import React from 'react'
-import { Text, View, StyleSheet } from 'react-native'
+import React, { useState, useEffect } from 'react'
+import { Text, View, StyleSheet, Button } from 'react-native'
+import swapi from '../api/swapi'
+import ItemCardList from '../components/ItemCardList'
 
 const ManageInventoryScreen = () => {
+    const [equipment, setEquipment] = useState([])
+
+    const searchApi = async ( limit ) => {
+        const response = await swapi.get('/equipment');
+        const limitArray = []
+        for (let i = 0; i < limit; i++) {
+            limitArray[i] = response.data[i]
+        }
+        //Currently the for loop is not being used and therefore the call is bringing in all data.
+        //This is by design at this time.
+        setEquipment(response.data)
+    }
+
+    useEffect(() => {
+        searchApi(20)},
+        [])
+
     return (
-        <View>
-            <Text>Manage Inventory Screen</Text>
+        <View style = { styles.screenContainer }>
+            <Text style = { styles.headerStyle }>Inventory Screen</Text>
+            <Text>Displaying {equipment.length} items.</Text>
+            {
+                equipment.length === 0
+                ? <Text>Equipment is Empty</Text>
+                : <ItemCardList
+                      equipment = { equipment }
+                  />
+            }
         </View>
     )
 }
 
-const styles = StyleSheet.create({})
+const styles = StyleSheet.create({
+    screenContainer: {
+        alignItems: 'center',
+        flex: 1
+    },
+    headerStyle: {
+        fontSize: 24,
+        fontWeight: 'bold',
+        alignSelf: 'center'
+    }
+})
 
 export default ManageInventoryScreen
